Update use to Ruby 2.0 and Rails 3.2 (minimal changes)
This is the minimal set of changes needed to make the application appear
to work on Heroku using Ruby 2.0 and Rails 3.2.

The major change is that we expect to compile the assets prior to pushing
to Heroku, so the precompilation needs to be done locally and the assets
tracked in git. To compile the gems:

    RAILS_ENV=production bundle exec rake assets:precompile

As `vendor/plugins` are depracated (directory removed) we can ignore
the entire `vendor/` tree and that lets me use `rbenv` and `bundler`
rather than `rvm` and a gemset for gem management. To install gems
the first time:

    bundle install --path vendor/bundle

Minimal changes to Gemfile to run with ruby 2.0.0 and rails 3.2

Changes to `config/environments/{development,test}.rb` and `script/rails` as
suggested at http://guides.rubyonrails.org/3_2_release_notes.html